Begin implementing instruction execution. This involves an execute method for the Avrcore. This fetches the instruction which the PC points at. Once fetched, .execute() is called on the instruction, which in turn executes the instruction and mutates the Avrcore as needed.
This commit includes the implementation of JMP.

+++ src/disassembler.rs
@@ -190,7 +190,7 @@ fn decode_jmp(opcode_words: Vec<u16>) -> JMPInstruction {
     let jmp_addr = ((top_5_bits | top_6_bit) as u32 | opcode_words[1] as u32)<<1;
 
     JMPInstruction {
-        address: jmp_addr
+        address: jmp_addr as usize
     }
 }
